# docs/swagger.yaml
definitions:
...
  errno.Errno:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  idea.IdeaResponse:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/model.IdeaListItem'
        type: array
    type: object
  message.GetMessageForCommentResponse:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/model.MessageForCommentItem'
        type: array
    type: object
  message.GetMessageForLikeResponse:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/model.MessageForLikeItem'
        type: array
    type: object
  message.GetMessageTipResponse:
    properties:
      is_message:
        type: boolean
    type: object
...
  model.IdeaListItem:
    properties:
      avatar:
        type: string
      comment_sum:
        type: integer
      content:
        type: string
      idea_id:
        type: integer
      liked:
        description: 是否点赞
        type: boolean
      likes_sum:
        type: integer
      nickname:
        type: string
      publisher_id:
        type: integer
      release_date:
        type: string
    type: object
  model.MessageForCommentItem:
    properties:
      avatar:
        type: string
      content:
        type: string
      date:
        type: string
      idea_id:
        type: integer
      nickname:
        type: string
      reply:
        type: string
      user_id:
        type: integer
    required:
    - avatar
    - content
    - idea_id
    - nickname
    - reply
    - user_id
    type: object
  model.MessageForLikeItem:
    properties:
      avatar:
        type: string
      comment_id:
        type: integer
      content:
        type: string
      date:
        type: string
      idea_id:
        type: integer
      nickname:
        type: string
      user_id:
        type: integer
    required:
    - avatar
    - comment_id
    - content
    - idea_id
    - nickname
    - user_id
    type: object
...
host: 124.71.184.107
info:
  contact:
    email: 2036479155@qq.com
    name: gongna
  description: 零重力APP
  termsOfService: http://swagger.io/terrms
  title: ZeroGravity
  version: 1.0.0
paths:
...
  /api/v1/idea/comment/:id/like:
    put:
      consumes:
      - application/json
      description: 新增评论点赞
      parameters:
      - description: LikersId点赞者的ID
        in: header
        name: token
        required: true
        type: string
      - description: CommentId 评论的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 上传对用户发布的评论的点赞
      tags:
      - idea
...
  /api/v1/idea/detail/:id/like:
    put:
      consumes:
      - application/json
      description: 新增对想法的点赞
      parameters:
      - description: 'LikersId点赞者的ID '
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 上传想法点赞信息
      tags:
      - idea
  /api/v1/idea/liked:
    get:
      consumes:
      - application/json
      description: 获取想法被点赞的信息表
      parameters:
      - description: UserId--用户ID
        in: header
        name: token
        required: true
        type: string
      - description: limit--偏移量指定开始返回记录之前要跳过的记录数
        in: query
        name: limit
        required: true
        type: integer
      - description: 'page--限制指定要检索的记录数 '
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/idea.IdeaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 获取用户个人发布的想法被点赞的情况
      tags:
      - idea
  /api/v1/idea/list:
    get:
      consumes:
      - application/json
      description: 显示用户发布的想法
      parameters:
      - description: uid--用户ID
        in: header
        name: token
        required: true
        type: string
      - description: limit--偏移量指定开始返回记录之前要跳过的记录数
        in: query
        name: limit
        required: true
        type: integer
      - description: 'page--限制指定要检索的记录数 '
        in: query
        name: page
        required: true
        type: integer
      - description: userId--用户ID
        in: query
        name: userId
        required: true
        type: integer
      - description: privicy 获取想法的私有策略 0->不获取私有 1->获取私有 默认不获取 || 若获取则默认为获取自己的想法，需要服务判断 uid
        in: query
        name: privicy
        required: true
        type: integer
      - description: index--获取排序规则 0->默认时间排序 1->热度排序
        in: query
        name: index
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/idea.IdeaResponse'
      summary: 进入空间后看到的想法
      tags:
      - idea
  /api/v1/message/comment:
    get:
      consumes:
      - application/json
      description: 拉取评论我的消息
      parameters:
      - description: 后端给前端的Token
        in: header
        name: token
        required: true
        type: string
      - description: limit--偏移量指定开始返回记录之前要跳过的记录数
        in: query
        name: limit
        required: true
        type: integer
      - description: page--限制指定要检索的记录数
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/message.GetMessageForCommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 拉取评论我的消息
      tags:
      - message
  /api/v1/message/like:
    get:
      consumes:
      - application/json
      description: 获取点赞我的信息
      parameters:
      - description: Like information returned to front-end users
        in: header
        name: token
        required: true
        type: string
      - description: limit--偏移量指定开始返回记录之前要跳过的记录数
        in: query
        name: limit
        required: true
        type: integer
      - description: page--限制指定要检索的记录数
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/message.GetMessageForLikeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 获取点赞我的信息
      tags:
      - message
  /api/v1/message/readall:
    put:
      consumes:
      - application/json
      description: 修改信息为已读
      parameters:
      - description: 后端给前端的token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 修改信息为已读
  /api/v1/message/tip:
    get:
      consumes:
      - application/json
      description: 获取消息提示
      parameters:
      - description: 后端给前端的Token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/message.GetMessageTipResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 获取消息提示
      tags:
      - message
...
schemes:
- http
swagger: "2.0"